pyproject.toml
[tool.poetry]
name = "diploma-staya"
version = "0.1.0"
description = "Autotesting demo project for staya.dog"
authors = ["Aleksandr Privalov <obojealexander@gmail.com>"]
readme = "README.md"
packages = [{include = "diploma_staya"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.2"
pytest = "^7.2.2"
voluptuous = "^0.13.1"
pytest-voluptuous = "^1.2.0"
allure-pytest = "^2.13.1"
python-dotenv = "^1.0.0"
curlify = "^2.2.1"
selene = "2.0.0b16"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
